+++ Model/ResourceModel/Entity.php
@@ -379,7 +379,7 @@ abstract class Entity extends \Magento\Framework\Model\ResourceModel\Db\Abstract
         if ($attribute->getIsRequired()) {
             // check if there is value of required attribute, its checked only on saving default values
             if ($object->getStoreId() == \Magento\Store\Model\Store::DEFAULT_STORE_ID) {
-                if ($value === null || $value == '' || (is_array($value) && empty($value))) {
+                if ($value === null || $value === '' || (is_array($value) && empty($value))) {
                     throw new \Exception(__('Value for attribute "%1" is required.', $attribute->getFrontendLabel()));
                 }
             }